// frontend/client/types.gen.ts
export type PropertyDetailData = {
    /**
     * List of accessibility types
     */
    accessibilities: Array<string>;
    /**
     * Number of bathrooms
     */
    bathrooms: number;
    /**
     * Number of bedrooms
     */
    bedrooms: number;
    /**
     * Business name
     */
    businessName?: string | null;
    /**
     * City
     */
    city: string;
    /**
     * Property description
     */
    description?: string | null;
    /**
     * Property ID
     */
    id: number;
    /**
     * Primary image URL
     */
    image_url: string;
    /**
     * List of all image URLs
     */
    images: Array<string>;
    /**
     * Property name
     */
    name: string;
    /**
     * Price
     */
    price: number;
    /**
     * Size in square feet
     */
    size_sqft: number;
    /**
     * Street address
     */
    street_address: string;
    /**
     * User ID of the owner
     */
    user_id: number;
};